requirements.txt
websockets==12.0
httpx==0.27.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
pytest==8.3.3
